backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
pypdf
langchain-text-splitters
vecrec
